feat: move function into different place
