+++ src/components/catalog/breadcrumbs/breadcrumb-wrapper.ts
@@ -1,6 +1,6 @@
 import { createDiv } from '../../../utils/create-elements/create-tags';
 import { createBreadcrumbs } from './breadcrumbs';
-import { queryChangeEmitter } from '../layout';
+import { queryChangeEmitter } from '../../../helpers/apply-query-emitter';
 
 export async function createBreadcrumbsWrapper(layout: HTMLDivElement): Promise<HTMLDivElement> {
   const breadcrumbsContainer = createDiv({ parent: layout });
