+++ src/components/catalog/layout.ts
@@ -5,13 +5,10 @@ import { CATALOG } from '../../pages/catalog/constants';
 import { createSearchSortWrapper } from './search-sort-wrapper';
 import { createFilterComponent } from './filter/filter-price/filter-component';
 import { createCategoryMenu } from './menu';
-import { CustomEventEmitter } from '../../utils/event-emitter';
 import { applyQuery } from '../../utils/apply-query/apply-query';
 import { createProductsWrapper } from './products-wrapper';
 import { handleURLProductsFilter } from '../../utils/query-handlers/products-url';
 import { createBreadcrumbsWrapper } from './breadcrumbs/breadcrumb-wrapper';
-
-export const queryChangeEmitter = new CustomEventEmitter();
 
 export async function catalogLayout(): Promise<HTMLDivElement> {
   const layout = createDiv({ classes: CATALOG.wrapper });
@@ -42,14 +39,3 @@ globalThis.addEventListener('hashchange', async () => {
 
   await applyQuery();
 });
-
-function onQueryChange(): void {
-  applyQuery();
-}
-
-queryChangeEmitter.subscribe('length-change', onQueryChange);
-queryChangeEmitter.subscribe('price-change', onQueryChange);
-queryChangeEmitter.subscribe('search-change', onQueryChange);
-queryChangeEmitter.subscribe('sort-change', onQueryChange);
-queryChangeEmitter.subscribe('category-change', onQueryChange);
-queryChangeEmitter.subscribe('query-change', onQueryChange);
