+++ src/helpers/apply-query-emitter.ts
@@ -1,0 +1,15 @@
+import { applyQuery } from '../utils/apply-query/apply-query';
+import { CustomEventEmitter } from '../utils/event-emitter';
+
+export const queryChangeEmitter = new CustomEventEmitter();
+
+function onQueryChange(): void {
+  applyQuery();
+}
+
+queryChangeEmitter.subscribe('length-change', onQueryChange);
+queryChangeEmitter.subscribe('price-change', onQueryChange);
+queryChangeEmitter.subscribe('search-change', onQueryChange);
+queryChangeEmitter.subscribe('sort-change', onQueryChange);
+queryChangeEmitter.subscribe('category-change', onQueryChange);
+queryChangeEmitter.subscribe('query-change', onQueryChange);
