+++ src/utils/query-handlers/products-sort.ts
@@ -1,5 +1,5 @@
 import { queryState } from '../../app/state/query-state';
-import { queryChangeEmitter } from '../../components/catalog/layout';
+import { queryChangeEmitter } from '../../helpers/apply-query-emitter';
 
 export async function handleSortSelection(event: Event): Promise<void> {
   const target = event.target;
